edupulse_demo/requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
selenium==4.17.2
google-generativeai>=0.8.0
ollama==0.6.1
python-dotenv==1.0.1
